added recursive token evaluation
Recursive token evaluation allows evaluated token to return other
tokens, which will then be re-evaluated in a recursive loop until there
are either no tokens remaining to evaluate, or the same tokens are
returned more than once (indicating they cannot be evaluated).

Additionally, I had to remove RegExpReplace functionality in
`EvalDialogTokens()` because the function would fail if the token's name
contained any characters that are reserved characters for regex,
including special characters like [.+^] etc.

Finally, I added `SetDialogTokenScript()`.  This function sets a
universal token evaluation function for any encountered token that was
not added *correctly* using AddDialogToken().  It does not allow for
that passage of a specific value set like AddDialogToken() does, but it
does allow for a universal token evaluation function that can evaluate
tokens that are not pre-defined.  If a token was added (and script
defined) with AddDialogToken(), that script will take priority over the
script set with `SetDialogTokenScript()`, so the functions can live
side-by-side without conflict.