feat: add recursive token evaluation and universal token eval script (#10)
* added recursive token evaluation

Recursive token evaluation allows evaluated tokens to return other tokens, which will then be re-evaluated in a recursive loop until there are either no tokens remaining to evaluate, or the same tokens are returned more than once (indicating they cannot be evaluated).

Additionally, I had to remove `RegExpReplace` functionality in `EvalDialogTokens()` because the function would fail if the token's name contained any characters that are reserved characters for regex, including special characters like [.+^] etc.

Finally, I added `SetDialogTokenScript()`.  This function sets a universal token evaluation function for any encountered token that was not added *correctly* using `AddDialogToken()`.  It does not allow for that passage of a specific value set like `AddDialogToken()` does, but it does allow for a universal token evaluation function that can evaluate
tokens that are not pre-defined.  If a token was added (and script defined) with `AddDialogToken()`, that script will take priority over the script set with `SetDialogTokenScript()`, so the functions can live side-by-side without conflict.

* allow use of double angle brackets to prevent token evaluation

I made the change to allow using `<<` and `>>` to ignore token evaluation and display the contained text within a single angle brackets. I think this is a good compromise since double-characters are often used in language to print a single characters. I also added a check so that any token that evaluates to itself is ignored, which should allow the use of an angle bracket as less-than or greather-than signs, but this might need more testing.